Структуры QuizQuestion, QuizStepViewModel и QuizResultsViewModel вынесены из контроллера в отдельный файл Models.swift
